# new_server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "new_server API",
    "version": "1.0.0",
    "description": "Machine-readable API contract for the Express backend in new_server."
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "apiKeyAuthScheme": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Alternative format: `Authorization: ApiKey <apiKey>`"
      },
      "inboundBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token for worker ingress"
      }
    },
    "schemas": {
      "RequestId": {
        "type": "string"
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "requestId"],
        "properties": {
          "error": { "type": "string" },
          "details": {},
          "requestId": { "$ref": "#/components/schemas/RequestId" }
        }
      },
...
      "EmailSender": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "name": { "type": "string", "nullable": true },
              "address": { "type": "string", "nullable": true }
            }
          },
          {
            "type": "null"
          }
        ]
      },
      "Email": {
        "type": "object",
        "required": ["id", "to", "localPart", "domain", "envelopeFrom", "from", "subject", "text", "html", "workerName", "sourceDomain", "messageId", "receivedAt", "createdAt", "hasRawMime", "rawMimeSize", "groupCount"],
        "properties": {
          "id": { "type": "integer" },
          "to": { "type": "string" },
          "localPart": { "type": "string" },
          "domain": { "type": "string" },
          "envelopeFrom": { "type": "string", "nullable": true },
          "from": { "$ref": "#/components/schemas/EmailSender" },
          "subject": { "type": "string" },
          "text": { "type": "string" },
          "html": { "type": "string" },
          "workerName": { "type": "string" },
          "sourceDomain": { "type": "string" },
          "messageId": { "type": "string" },
          "receivedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "hasRawMime": { "type": "boolean" },
          "rawMimeSize": { "type": "integer" },
          "groupCount": { "type": "integer" },
          "rawMime": { "type": "string", "nullable": true }
        }
      },
...
      "GroupOwner": {
        "oneOf": [
          {
            "type": "object",
            "required": ["id", "username", "displayName"],
            "properties": {
              "id": { "type": "integer" },
              "username": { "type": "string" },
              "displayName": { "type": "string" }
            }
          },
          {
            "type": "null"
          }
        ]
      },
      "Group": {
        "type": "object",
        "required": ["id", "ownerUserId", "owner", "name", "color", "description", "emailCount", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "integer" },
          "ownerUserId": { "type": "integer" },
          "owner": { "$ref": "#/components/schemas/GroupOwner" },
          "name": { "type": "string" },
          "color": { "type": "string" },
          "description": { "type": "string" },
          "emailCount": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "GroupEmail": {
        "allOf": [
          { "$ref": "#/components/schemas/Email" },
          {
            "type": "object",
            "required": ["groupPosition", "groupAddedAt", "groupAddedByUserId"],
            "properties": {
              "groupPosition": { "type": "integer" },
              "groupAddedAt": { "type": "string", "format": "date-time" },
              "groupAddedByUserId": { "type": "integer", "nullable": true }
            }
          }
        ]
      },
...
      "GroupRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "color": { "type": "string" },
          "description": { "type": "string" }
        }
      },
      "GroupEmailBatchRequest": {
        "type": "object",
        "required": ["emailIds"],
        "properties": {
          "emailIds": {
            "type": "array",
            "maxItems": 200,
            "items": {
              "oneOf": [{ "type": "integer" }, { "type": "string" }]
            }
          },
          "includeRawMime": { "type": "boolean" }
        }
      },
      "SimpleSuccessResponse": {
        "type": "object",
        "required": ["success", "requestId"],
        "properties": {
          "success": { "type": "boolean" },
          "requestId": { "$ref": "#/components/schemas/RequestId" }
        }
      }
    },
    "responses": {
...
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/groups": {
      "get": {
        "tags": ["Groups"],
        "summary": "List groups for current user or ownerUserId for admins",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "query", "name": "ownerUserId", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Group list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "groups": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Group" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Groups"],
        "summary": "Create group",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GroupRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "group": { "$ref": "#/components/schemas/Group" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/v1/groups/{groupId}": {
      "get": {
        "tags": ["Groups"],
        "summary": "Get group",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "groupId", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Group details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "group": { "$ref": "#/components/schemas/Group" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Groups"],
        "summary": "Update group",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "groupId", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GroupRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "group": { "$ref": "#/components/schemas/Group" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      },
      "delete": {
        "tags": ["Groups"],
        "summary": "Delete group",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "groupId", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Group deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SimpleSuccessResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/groups/{groupId}/emails": {
      "get": {
        "tags": ["Groups"],
        "summary": "List group emails",
        "description": "Group visibility is evaluated against the group owner's mailbox registrations and permissions. If any saved email id is now missing or denied, the backend removes it from the group and returns 409 so the client can reload.",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "groupId", "required": true, "schema": { "type": "integer" } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 100, "maximum": 200 } },
          { "in": "query", "name": "offset", "schema": { "type": "integer", "default": 0 } },
          { "in": "query", "name": "includeRawMime", "schema": { "type": "string", "enum": ["1"] } }
        ],
        "responses": {
          "200": {
            "description": "Group emails",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "group": { "$ref": "#/components/schemas/Group" },
                    "total": { "type": "integer" },
                    "count": { "type": "integer" },
                    "emails": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/GroupEmail" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      },
      "post": {
        "tags": ["Groups"],
        "summary": "Append emails to group",
        "description": "A group is permanently owned by one user. Every emailId added must resolve to a mailbox registered by that owner, otherwise the request is rejected.",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "groupId", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GroupEmailBatchRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Emails appended to group",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "group": { "$ref": "#/components/schemas/Group" },
                    "total": { "type": "integer" },
                    "count": { "type": "integer" },
                    "emails": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/GroupEmail" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Groups"],
        "summary": "Replace group emails",
        "description": "Replaces the entire group membership. Every emailId must still be visible to the group owner through both permission scope and mailbox registration.",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "groupId", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GroupEmailBatchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group emails replaced",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "group": { "$ref": "#/components/schemas/Group" },
                    "total": { "type": "integer" },
                    "count": { "type": "integer" },
                    "emails": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/GroupEmail" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/groups/{groupId}/emails/{emailId}": {
      "delete": {
        "tags": ["Groups"],
        "summary": "Remove one email from group",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "groupId", "required": true, "schema": { "type": "integer" } },
          { "in": "path", "name": "emailId", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Email removed from group",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SimpleSuccessResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}